some general cleanup, consolidated Detail and DetailsTable components into RestaurantInfo file, and RatingStars into RatingsTable
